+++ modules/elements/elements/hd-flipcard/templates/template-link.php
@@ -1,6 +1,6 @@
 <?php
 
-/* Herzog Dupont for YOOtheme Pro Copyright (C) 2016-2023 YOOtheme GmbH, 2019-2023 Thomas Weidlich GNU GPL v3 */
+/* Herzog Dupont for YOOtheme Pro Copyright (C) 2016-2024 YOOtheme GmbH, 2019-2024 Thomas Weidlich GNU GPL v3 */
 
 $link = $props['link_text'] ? $this->el('p', [
     //