+++ modules/elements/elements/hd-image-comparison/js/hd-image-comparison.js
@@ -1,4 +1,4 @@
-/* Herzog Dupont for YOOtheme Pro Copyright (C) 2020-2023 Thomas Weidlich GNU GPL v3 */
+/* Herzog Dupont for YOOtheme Pro Copyright (C) 2020-2024 Thomas Weidlich GNU GPL v3 */
 
 class hdImgComp {