+++ modules/elements/elements/hd-progress/templates/content.php
@@ -1,6 +1,6 @@
 <?php
 
-/* Herzog Dupont for YOOtheme Pro Copyright (C) 2018-2023 Thomas Weidlich GNU GPL v3 */
+/* Herzog Dupont for YOOtheme Pro Copyright (C) 2018-2024 Thomas Weidlich GNU GPL v3 */
 
 if ($props['content']) : ?>
 <div><?= $props['content'] ?></div>