+++ script.php
@@ -3,7 +3,7 @@
 /**
  * @package   Herzog Dupont for YOOtheme Pro
  * @author    Thomas Weidlich
- * @copyright Copyright (C) 2021-2023 Thomas Weidlich
+ * @copyright Copyright (C) 2021-2024 Thomas Weidlich
  * @license   GNU General Public License version 3, see LICENSE
  */
 
@@ -14,7 +14,7 @@ class plgSystemHerzogdupontInstallerScript
 {
     protected $minimumPHPVersion = '8.0.0';
     protected $minimumJoomlaVersion = '4.0.0';
-    protected $minimumYOOthemeVersion = '4.0.0';
+    protected $minimumYOOthemeVersion = '4.3.0';
 
     public function install($parent)
     {